firebase, flipcard, svgr, tailwind3d, route
Route" configuration for navigation bar, initialization of route index, addition of sign in and sign out, addition of flip card for sign in and sign out page, addition of tailwindcss 3d package and svgr rollup, addition of theme extension for rotate and boxshadow in tailwindcss and integration of tailwindcss-3d plugin, integration of svgr in vite configuration for transformation of logo into reusable component, addition of firebase package not yet integrated. correction of card and flipcard style with shadow on tailwind.